+++ internal/cmd/self-update.go
@@ -20,14 +20,14 @@ var selfUpdateCommand = &cobra.Command{
 		// TODO
 		// call public github api
 		// if latest version is not current version (does not match)
-		// then download get.sh
+		// then download install.sh
 
-		fileUrl := "https://raw.githubusercontent.com/craftcms/nitro/master/get.sh"
+		fileUrl := "https://raw.githubusercontent.com/craftcms/nitro/master/install.sh"
 
 		tempFolder := os.TempDir()
 		defer os.Remove(tempFolder)
 
-		localFile := filepath.Join(tempFolder, "get.sh")
+		localFile := filepath.Join(tempFolder, "install.sh")
 
 		if err := DownloadFile(localFile, fileUrl); err != nil {
 			return err